Partial testing of upload completed:
