--- Zootils/__init__.py
@@ -1,5 +1,0 @@
-# NEEDS to be tested
-from JamesLab.Zootils import utils
-from JamesLab.Zootils import upload
-from JamesLab.Zootils import manifest
-from JamesLab.Zootils import export
