--- Zootils/__main__.py
@@ -1,4 +1,0 @@
-from JamesLab.apps.base import dmain
-
-if __name__ == '__main__':
-    dmain(__file__)
