--- Zootils/test/img/manifest.csv
@@ -1,9 +1,0 @@
-id,filename
-122118-194845-0000,img\282_CM001__2017-08-13__False (2).jpg
-122118-194845-0001,img\282_CM001__2017-08-13__False (3).jpg
-122118-194845-0002,img\282_CM001__2017-08-13__False (4).jpg
-122118-194845-0003,img\282_CM001__2017-08-13__False (5).jpg
-122118-194845-0004,img\282_CM001__2017-08-13__False (6).jpg
-122118-194845-0005,img\282_CM001__2017-08-13__False (7).jpg
-122118-194845-0006,img\282_CM001__2017-08-13__False (8).jpg
-122118-194845-0007,img\282_CM001__2017-08-13__False.jpg
