+++ Zootilsdir/__init__.py
@@ -1,0 +1,5 @@
+# NEEDS to be tested
+from JamesLab.Zootils import utils
+from JamesLab.Zootils import upload
+from JamesLab.Zootils import manifest
+from JamesLab.Zootils import export
