+++ Zootilsdir/__main__.py
@@ -1,0 +1,4 @@
+from JamesLab.apps.base import dmain
+
+if __name__ == '__main__':
+    dmain(__file__)
