+++ Zootilsdir/test/JamesLab.Zootils.log
@@ -1,0 +1,3 @@
+Creating new subject set
+Creating new subject set
+Creating new subject set
