+++ Zootilsdir/test/img/manifest.csv
@@ -1,0 +1,17 @@
+id,filename
+122218-103723-0000,D:\SchnableLab\JamesLab\Zootilsdir\test\img\282_CM001__2017-08-13__False (2).jpg
+122218-103723-0001,D:\SchnableLab\JamesLab\Zootilsdir\test\img\282_CM001__2017-08-13__False (3).jpg
+122218-103723-0002,D:\SchnableLab\JamesLab\Zootilsdir\test\img\282_CM001__2017-08-13__False (4).jpg
+122218-103723-0003,D:\SchnableLab\JamesLab\Zootilsdir\test\img\282_CM001__2017-08-13__False (5).jpg
+122218-103723-0004,D:\SchnableLab\JamesLab\Zootilsdir\test\img\282_CM001__2017-08-13__False (6).jpg
+122218-103723-0005,D:\SchnableLab\JamesLab\Zootilsdir\test\img\282_CM001__2017-08-13__False (7).jpg
+122218-103723-0006,D:\SchnableLab\JamesLab\Zootilsdir\test\img\282_CM001__2017-08-13__False (8).jpg
+122218-103723-0007,D:\SchnableLab\JamesLab\Zootilsdir\test\img\282_CM001__2017-08-13__False.jpg
+122218-103723-0008,D:\SchnableLab\JamesLab\Zootilsdir\test\img\282_CM001__2017-08-13__False.png
+122218-103723-0009,D:\SchnableLab\JamesLab\Zootilsdir\test\img\284_CM003__2017-08-19__True.png
+122218-103723-0010,D:\SchnableLab\JamesLab\Zootilsdir\test\img\285_CM004__2017-08-16__True.png
+122218-103723-0011,D:\SchnableLab\JamesLab\Zootilsdir\test\img\285_CM004__2017-08-23__True.png
+122218-103723-0012,D:\SchnableLab\JamesLab\Zootilsdir\test\img\288_CM006__2017-08-27__True.png
+122218-103723-0013,D:\SchnableLab\JamesLab\Zootilsdir\test\img\289_CM007__2017-08-21__True.png
+122218-103723-0014,D:\SchnableLab\JamesLab\Zootilsdir\test\img\289_CM007__2017-08-27__True.png
+122218-103723-0015,D:\SchnableLab\JamesLab\Zootilsdir\test\img\290_CM008__2017-08-01__False.png
